fix shellexecution deleting already inactive unit

+++ src/core/execution/ShellExecution.ts
@@ -20,6 +20,10 @@ export class ShellExecution implements Execution {
         if (this.shell == null) {
             this.shell = this._owner.buildUnit(UnitType.Shell, 0, this.spawn)
         }
+        if (!this.shell.isActive()) {
+            this.active = false
+            return
+        }
         if (!this.target.isActive()) {
             this.shell.delete()
             this.active = false
